Journal members: don't allow users adding themselves as journal members.
This is not a big deal, but it's just weird that you can give yourself
access to a journal you already have access to.
It currently doesn't show a nice error message, but at least it prevents
the change from being applied.

This fixes:
https://github.com/etesync/android/issues/11

The issue was reported on the client repository, but it's actually a
server validation issue.