Implement journal attachments.
Journal attachments are essentially the same as the existing journal
entries. The only difference is that attachments are mutable, and
can be deleted while entries can't.
This is on the server side. On the client side, there's also a slight
difference in the encryption scheme. Attachments are not chained, but
rather are all direct children of the journal. So they should all be
encrypted with prevUid set to null.

Attachments are useful for keeping larger files adjacent to the journal
which opens a few interesting possibilities:
1. You can store larger files in adjacent to the journal and use the
journal to refer to them, so you can be certain they are new, but still
not storing history
2. Same as the first point, but instead of larger files, just files that
you'd like to only keep the most recent version of but not maintain its
history. Think for example GPS location.
3. Slightly related to the previous point, you can use attachments on
their own, without referring to them from the journal for sharing data
temporarily. For example to share a file securely.