Fix race condition when adding entries to journal.
While rare, an issue with a client could make it so two concurrent
requests (POST) to add two different entries but passing the same "last"
parameter would both pass the last validation and therefore both get
inserted. This however would be wrong as the second to be inserted will
have the wrong integrity validation (as it was calculated against last
rather than against the current previous entry).

This is not theoretical, people seem to have managed to trigger this.
Especially with tasks.