Really fix race condition when adding entries to journal.
More details are available in 891fe2a1ec6d0b510d019cef8a2671d7755e3707
What was wrong with that fix though, is that after the lock we would
always get the previous last entry, which is obviously wrong. With this
fix we now fetch the correct entry after the lock has been freed.